requirements.txt
requests
psycopg2-binary
pandas>=1.3.0
